+++ whywhyanalysisApp/AnalysisListCustumCell.xib
@@ -9,7 +9,7 @@
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="analysisListCustumCell" customModule="whywhyanalysisApp" customModuleProvider="target"/>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view clearsContextBeforeDrawing="NO" contentMode="scaleToFill" id="i5M-Pr-FkT" customClass="analysisListCustumCell" customModule="whywhyanalysisApp" customModuleProvider="target">
+        <view clearsContextBeforeDrawing="NO" contentMode="scaleToFill" id="i5M-Pr-FkT" customClass="AnalysisListCustumCell" customModule="whywhyanalysisApp" customModuleProvider="target">
             <rect key="frame" x="0.0" y="0.0" width="320" height="568"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
@@ -31,8 +31,8 @@
             <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
             <viewLayoutGuide key="safeArea" id="fnl-2z-Ty3"/>
             <connections>
-                <outlet property="measuresLabel" destination="Yc4-at-u87" id="KsE-sD-tiY"/>
-                <outlet property="problemLabel" destination="2NN-sQ-tG0" id="iBp-F7-JW2"/>
+                <outlet property="measuresLabel" destination="Yc4-at-u87" id="qol-Fn-AGm"/>
+                <outlet property="problemLabel" destination="2NN-sQ-tG0" id="yzc-hC-1z1"/>
             </connections>
             <point key="canvasLocation" x="-258.75" y="66.549295774647888"/>
         </view>
