+++ whywhyanalysisApp/Base.lproj/Main.storyboard
@@ -26,7 +26,7 @@
         <!--Home View Controller-->
         <scene sceneID="RCg-FY-SAZ">
             <objects>
-                <viewController storyboardIdentifier="top" id="i5Z-bF-1Z8" customClass="HomeViewController" customModule="whywhyanalysisApp" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="top" id="i5Z-bF-1Z8" customClass="HomeViewController" customModule="whywhyanalysisApp" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="ggm-Yj-bg3">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
