+++ whywhyanalysisApp/Model/Whywhyanalysis.swift
