+++ whywhyanalysisApp/Utility/ValidateUtility.swift
